+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [0.101.0] - 2026-02-03
+
+### Changes
+- Add location filtering to regular volunteers admin page ([#546](https://github.com/everybody-eats-nz/volunteer-portal/pull/546)) by @malinmalliyawadu
+
+
+## [0.100.3] - 2026-02-01
+
+### Changes
+- chore(deps): bump posthog-js from 1.335.2 to 1.336.4 in /web ([#540](https://github.com/everybody-eats-nz/volunteer-portal/pull/540)) by @dependabot[bot]
+
+
 ## [0.100.2] - 2026-02-01
 
 ### Changes